[Ref]: nowPlaying handler 코드 분리
